requirements.txt
black==18.9b0
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.7
django-debug-toolbar==1.11
django-dotenv==1.4.2
django-extensions==2.0.7
django-widget-tweaks==1.4.2
flake8==3.5.0
gunicorn==19.6.0
idna==2.7
Markdown==3.0.1
mccabe==0.6.1
oauthlib==2.1.0
pre-commit==1.14.2
psycopg2-binary==2.7.4
pycodestyle==2.3.1
pyflakes==1.6.0
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.3
requests==2.20.0
requests-oauthlib==1.0.0
sentry-sdk==0.6.9
shortuuid==0.5.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
trueskill==0.4.5
urllib3==1.23
